requirements.txt
torch==2.1.2
transformers==4.33.3
glove-python-binary==0.2.0; python_version < "3.9"
tqdm==4.66.5
pyyaml==6.0.1
wandb==0.17.9
huggingface_hub==0.34.4
